@@ -1,56 +1,85 @@
 [package]
-name = "exa"
+name = "eza"
 description = "A modern replacement for ls"
-
-authors = ["Benjamin Sago <ogham@bsago.me>"]
+authors = ["Christina Sørensen <christina@cafkafk.com>"]
 categories = ["command-line-utilities"]
-edition = "2018"
+edition = "2021"
+rust-version = "1.65.0"
 exclude = ["/devtools/*", "/Justfile", "/Vagrantfile", "/screenshots.png"]
-homepage = "https://the.exa.website/"
+readme = "README.md"
+homepage = "https://github.com/eza-community/eza"
 license = "MIT"
-repository = "https://github.com/ogham/exa"
-version = "0.10.1"
+repository = "https://github.com/eza-community/eza"
+version = "0.11.1"
+
+
+[package.metadata.deb]
+license-file = [ "LICENCE", "4" ]
+depends = "$auto"
+extended-description = """
+eza is a modern, maintained replacement for ls
+"""
+section = "utils"
+priority = "optional"
+assets = [
+    [ "target/release/eza", "/usr/bin/eza", "0755" ],
+    [ "target/release/../man/eza.1", "/usr/share/man/man1/eza.1", "0644" ],
+    [ "target/release/../man/eza_colors.5", "/usr/share/man/man5/eza_colors.5", "0644" ],
+    [ "completions/bash/eza", "/usr/share/bash-completion/completions/eza", "0644" ],
+    [ "completions/zsh/_eza", "/usr/share/zsh/site-functions/_eza", "0644" ],
+    [ "completions/fish/eza.fish", "/usr/share/fish/vendor_completions.d/eza.fish", "0644" ],
+]
 
 
 [[bin]]
-name = "exa"
+name = "eza"
 
 
 [dependencies]
-ansi_term = "0.12"
+ansiterm = "0.12.2"
+gethostname = "0.4.3"
+chrono = { version = "0.4.30", default-features = false, features = ["clock"] }
 glob = "0.3"
 lazy_static = "1.3"
 libc = "0.2"
 locale = "0.2"
 log = "0.4"
 natord = "1.0"
-num_cpus = "1.10"
+num_cpus = "1.16"
 number_prefix = "0.4"
+phf = { version = "0.11.2", features = ["macros"]}
 scoped_threadpool = "0.1"
 term_grid = "0.1"
-terminal_size = "0.1.16"
+terminal_size = "0.2.6"
+timeago = { version = "0.4.1", default-features = false }
 unicode-width = "0.1"
-users = "0.11"
+urlencoding = "2.1.3"
 zoneinfo_compiled = "0.5.1"
 
-[dependencies.datetime]
-version = "0.5.2"
-default-features = false
-features = ["format"]
-
 [dependencies.git2]
-version = "0.13"
+version = "0.18"
 optional = true
 default-features = false
 
-[build-dependencies.datetime]
-version = "0.5.2"
-default-features = false
+[target.'cfg(target_os = "linux")'.dependencies]
+proc-mounts = "0.3"
+
+[target.'cfg(unix)'.dependencies]
+uzers = "0.11.3"
+
+[build-dependencies]
+chrono = { version = "0.4.30", default-features = false, features = ["clock"] }
+
+[dev-dependencies]
+trycmd = "0.14"
 
 [features]
 default = [ "git" ]
 git = [ "git2" ]
 vendored-openssl = ["git2/vendored-openssl"]
+vendored-libgit2 = ["git2/vendored-libgit2"]
+# Should only be used inside of flake.nix
+nix = []
 
 
 # make dev builds faster by excluding debug symbols
@@ -60,18 +89,5 @@ debug = false
 # use LTO for smaller binaries (that take longer to build)
 [profile.release]
 lto = true
-
-
-[package.metadata.deb]
-license-file = [ "LICENCE" ]
-depends = "$auto"
-extended-description = """
-exa is a replacement for ls written in Rust.
-"""
-section = "utils"
-priority = "optional"
-assets = [
-    [ "target/release/exa", "/usr/bin/exa", "0755" ],
-    [ "contrib/man/exa.1", "/usr/share/man/man1/exa.1", "0644" ],
-    [ "contrib/completions.bash", "/etc/bash_completion.d/exa", "0644" ],
-]
+strip = true
+opt-level = "s"
