@@ -6,11 +6,11 @@ use crate::options::parser::MatchedFlags;
 
 
 static USAGE_PART1: &str = "Usage:
-  exa [options] [files...]
+  eza [options] [files...]
 
 META OPTIONS
   -?, --help         show list of command-line options
-  -v, --version      show version of exa
+  -v, --version      show version of eza
 
 DISPLAY OPTIONS
   -1, --oneline      display one entry per line
@@ -24,9 +24,11 @@ DISPLAY OPTIONS
   --colo[u]r-scale   highlight levels of file sizes distinctly
   --icons            display icons
   --no-icons         don't display icons (always overrides --icons)
+  --hyperlink        display entries as hyperlinks
+  -w, --width COLS   set screen width in columns
 
 FILTERING AND SORTING OPTIONS
-  -a, --all                  show hidden and 'dot' files
+  -a, --all                  show hidden and 'dot' files. Use this twice to also show the '.' and '..' directories
   -d, --list-dirs            list directories as files; don't list their contents
   -L, --level DEPTH          limit the depth of recursion
   -r, --reverse              reverse the sort order
@@ -35,41 +37,49 @@ FILTERING AND SORTING OPTIONS
   -D, --only-dirs            list only directories
   -I, --ignore-glob GLOBS    glob patterns (pipe-separated) of files to ignore";
 
-  static USAGE_PART2: &str = "  \
+static GIT_FILTER_HELP: &str = "  \
+  --git-ignore               ignore files mentioned in '.gitignore'";
+
+static USAGE_PART2: &str = "  \
   Valid sort fields:         name, Name, extension, Extension, size, type,
                              modified, accessed, created, inode, and none.
                              date, time, old, and new all refer to modified.
 
 LONG VIEW OPTIONS
-  -b, --binary         list file sizes with binary prefixes
-  -B, --bytes          list file sizes in bytes, without any prefixes
-  -g, --group          list each file's group
-  -h, --header         add a header row to each column
-  -H, --links          list each file's number of hard links
-  -i, --inode          list each file's inode number
-  -m, --modified       use the modified timestamp field
-  -n, --numeric        list numeric user and group IDs
-  -S, --blocks         show number of file system blocks
-  -t, --time FIELD     which timestamp field to list (modified, accessed, created)
-  -u, --accessed       use the accessed timestamp field
-  -U, --created        use the created timestamp field
-  --changed            use the changed timestamp field
-  --time-style         how to format timestamps (default, iso, long-iso, full-iso)
-  --no-permissions     suppress the permissions field
-  --octal-permissions  list each file's permission in octal format
-  --no-filesize        suppress the filesize field
-  --no-user            suppress the user field
-  --no-time            suppress the time field";
+  -b, --binary             list file sizes with binary prefixes
+  -B, --bytes              list file sizes in bytes, without any prefixes
+  -g, --group              list each file's group
+  -h, --header             add a header row to each column
+  -H, --links              list each file's number of hard links
+  -i, --inode              list each file's inode number
+  -m, --modified           use the modified timestamp field
+  -M, --mounts             show mount details (Linux only)
+  -n, --numeric            list numeric user and group IDs
+  -S, --blocksize          show size of allocated file system blocks
+  -t, --time FIELD         which timestamp field to list (modified, accessed, created)
+  -u, --accessed           use the accessed timestamp field
+  -U, --created            use the created timestamp field
+  --changed                use the changed timestamp field
+  --time-style             how to format timestamps (default, iso, long-iso, full-iso, relative)
+  --no-permissions         suppress the permissions field
+  -o, --octal-permissions  list each file's permission in octal format
+  --no-filesize            suppress the filesize field
+  --no-user                suppress the user field
+  --no-time                suppress the time field";
 
-static GIT_FILTER_HELP: &str = "  --git-ignore               ignore files mentioned in '.gitignore'";
-static GIT_VIEW_HELP:   &str = "  --git                list each file's Git status, if tracked or ignored";
-static EXTENDED_HELP:   &str = "  -@, --extended       list each file's extended attributes and sizes";
-
+static GIT_VIEW_HELP:   &str = "  \
+  --git                    list each file's Git status, if tracked or ignored
+  --no-git                 suppress Git status (always overrides --git, --git-repos, --git-repos-no-status)
+  --git-repos              list root of git-tree status";
+static EXTENDED_HELP:   &str = "  \
+  -@, --extended           list each file's extended attributes and sizes";
+static SECATTR_HELP:    &str = "  \
+  -Z, --context            list each file's security context";
 
 /// All the information needed to display the help text, which depends
 /// on which features are enabled and whether the user only wants to
 /// see one section’s help.
-#[derive(PartialEq, Debug, Copy, Clone)]
+#[derive(PartialEq, Eq, Debug, Copy, Clone)]
 pub struct HelpString;
 
 impl HelpString {
@@ -96,20 +106,21 @@ impl fmt::Display for HelpString {
     /// Format this help options into an actual string of help
     /// text to be displayed to the user.
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        write!(f, "{}", USAGE_PART1)?;
+        write!(f, "{USAGE_PART1}")?;
 
         if cfg!(feature = "git") {
-            write!(f, "\n{}", GIT_FILTER_HELP)?;
+            write!(f, "\n{GIT_FILTER_HELP}")?;
         }
 
-        write!(f, "\n{}", USAGE_PART2)?;
+        write!(f, "\n{USAGE_PART2}")?;
 
         if cfg!(feature = "git") {
-            write!(f, "\n{}", GIT_VIEW_HELP)?;
+            write!(f, "\n{GIT_VIEW_HELP}")?;
         }
 
         if xattr::ENABLED {
-            write!(f, "\n{}", EXTENDED_HELP)?;
+            write!(f, "\n{EXTENDED_HELP}")?;
+            write!(f, "\n{SECATTR_HELP}")?;
         }
 
         writeln!(f)
