@@ -216,7 +216,7 @@ pub mod test {
     use crate::options::parser::{Arg, MatchedFlags};
     use std::ffi::OsStr;
 
-    #[derive(PartialEq, Debug)]
+    #[derive(PartialEq, Eq, Debug)]
     pub enum Strictnesses {
         Last,
         Complain,
@@ -228,14 +228,14 @@ pub mod test {
     /// both, then both should resolve to the same result.
     ///
     /// It returns a vector with one or two elements in.
-    /// These elements can then be tested with assert_eq or what have you.
+    /// These elements can then be tested with `assert_eq` or what have you.
     pub fn parse_for_test<T, F>(inputs: &[&str], args: &'static [&'static Arg], strictnesses: Strictnesses, get: F) -> Vec<T>
     where F: Fn(&MatchedFlags<'_>) -> T
     {
         use self::Strictnesses::*;
         use crate::options::parser::{Args, Strictness};
 
-        let bits = inputs.into_iter().map(OsStr::new).collect::<Vec<_>>();
+        let bits = inputs.iter().map(OsStr::new).collect::<Vec<_>>();
         let mut result = Vec::new();
 
         if strictnesses == Last || strictnesses == Both {
