@@ -22,11 +22,12 @@ pub struct View {
     pub mode: Mode,
     pub width: TerminalWidth,
     pub file_style: file_name::Options,
+    pub deref_links: bool,
 }
 
 
 /// The **mode** is the “type” of output.
-#[derive(PartialEq, Debug)]
+#[derive(PartialEq, Eq, Debug)]
 #[allow(clippy::large_enum_variant)]
 pub enum Mode {
     Grid(grid::Options),
@@ -37,7 +38,7 @@ pub enum Mode {
 
 
 /// The width of the terminal requested by the user.
-#[derive(PartialEq, Debug, Copy, Clone)]
+#[derive(PartialEq, Eq, Debug, Copy, Clone)]
 pub enum TerminalWidth {
 
     /// The user requested this specific number of columns.
