@@ -1,21 +1,28 @@
-use ansi_term::Style;
-use users::{Users, Groups};
+use ansiterm::Style;
+use uzers::{Users, Groups};
 
 use crate::fs::fields as f;
 use crate::output::cell::TextCell;
 use crate::output::table::UserFormat;
 
+pub trait Render{
+    fn render<C: Colours, U: Users+Groups>(self, colours: &C, users: &U, format: UserFormat) -> TextCell;
+}
 
-impl f::Group {
-    pub fn render<C: Colours, U: Users+Groups>(self, colours: &C, users: &U, format: UserFormat) -> TextCell {
-        use users::os::unix::GroupExt;
+impl Render for Option<f::Group> {
+    fn render<C: Colours, U: Users+Groups>(self, colours: &C, users: &U, format: UserFormat) -> TextCell {
+        use uzers::os::unix::GroupExt;
 
         let mut style = colours.not_yours();
 
-        let group = match users.get_group_by_gid(self.0) {
-            Some(g)  => (*g).clone(),
-            None     => return TextCell::paint(style, self.0.to_string()),
+        let group = match self {
+            Some(g) => match users.get_group_by_gid(g.0) {
+                Some(g) => (*g).clone(),
+                None    => return TextCell::paint(style, g.0.to_string()),
+            },
+            None => return TextCell::blank(colours.no_group()),
         };
+
 
         let current_uid = users.get_current_uid();
         if let Some(current_user) = users.get_user_by_uid(current_uid) {
@@ -40,22 +47,23 @@ impl f::Group {
 pub trait Colours {
     fn yours(&self) -> Style;
     fn not_yours(&self) -> Style;
+    fn no_group(&self) -> Style;
 }
 
 
 #[cfg(test)]
 #[allow(unused_results)]
 pub mod test {
-    use super::Colours;
+    use super::{Colours, Render};
     use crate::fs::fields as f;
     use crate::output::cell::TextCell;
     use crate::output::table::UserFormat;
 
-    use users::{User, Group};
-    use users::mock::MockUsers;
-    use users::os::unix::GroupExt;
-    use ansi_term::Colour::*;
-    use ansi_term::Style;
+    use uzers::{User, Group};
+    use uzers::mock::MockUsers;
+    use uzers::os::unix::GroupExt;
+    use ansiterm::Colour::*;
+    use ansiterm::Style;
 
 
     struct TestColours;
@@ -63,6 +71,7 @@ pub mod test {
     impl Colours for TestColours {
         fn yours(&self)     -> Style { Fixed(80).normal() }
         fn not_yours(&self) -> Style { Fixed(81).normal() }
+        fn no_group(&self)   -> Style { Black.italic() }
     }
 
 
@@ -71,7 +80,7 @@ pub mod test {
         let mut users = MockUsers::with_current_uid(1000);
         users.add_group(Group::new(100, "folk"));
 
-        let group = f::Group(100);
+        let group = Some(f::Group(100));
         let expected = TextCell::paint_str(Fixed(81).normal(), "folk");
         assert_eq!(expected, group.render(&TestColours, &users, UserFormat::Name));
 
@@ -84,7 +93,7 @@ pub mod test {
     fn unnamed() {
         let users = MockUsers::with_current_uid(1000);
 
-        let group = f::Group(100);
+        let group = Some(f::Group(100));
         let expected = TextCell::paint_str(Fixed(81).normal(), "100");
         assert_eq!(expected, group.render(&TestColours, &users, UserFormat::Name));
         assert_eq!(expected, group.render(&TestColours, &users, UserFormat::Numeric));
@@ -96,7 +105,7 @@ pub mod test {
         users.add_user(User::new(2, "eve", 100));
         users.add_group(Group::new(100, "folk"));
 
-        let group = f::Group(100);
+        let group = Some(f::Group(100));
         let expected = TextCell::paint_str(Fixed(80).normal(), "folk");
         assert_eq!(expected, group.render(&TestColours, &users, UserFormat::Name))
     }
@@ -109,14 +118,14 @@ pub mod test {
         let test_group = Group::new(100, "folk").add_member("eve");
         users.add_group(test_group);
 
-        let group = f::Group(100);
+        let group = Some(f::Group(100));
         let expected = TextCell::paint_str(Fixed(80).normal(), "folk");
         assert_eq!(expected, group.render(&TestColours, &users, UserFormat::Name))
     }
 
     #[test]
     fn overflow() {
-        let group = f::Group(2_147_483_648);
+        let group = Some(f::Group(2_147_483_648));
         let expected = TextCell::paint_str(Fixed(81).normal(), "2147483648");
         assert_eq!(expected, group.render(&TestColours, &MockUsers::with_current_uid(0), UserFormat::Numeric));
     }
