@@ -1,4 +1,6 @@
+#[cfg(unix)]
 mod blocks;
+#[cfg(unix)]
 pub use self::blocks::Colours as BlocksColours;
 
 mod filetype;
@@ -7,9 +9,12 @@ pub use self::filetype::Colours as FiletypeColours;
 mod git;
 pub use self::git::Colours as GitColours;
 
+#[cfg(unix)]
 mod groups;
-pub use self::groups::Colours as GroupColours;
+#[cfg(unix)]
+pub use self::groups::{Colours as GroupColours, Render as GroupRender};
 
+#[cfg(unix)]
 mod inode;
 // inode uses just one colour
 
@@ -17,7 +22,7 @@ mod links;
 pub use self::links::Colours as LinksColours;
 
 mod permissions;
-pub use self::permissions::Colours as PermissionsColours;
+pub use self::permissions::{Colours as PermissionsColours, PermissionsPlusRender};
 
 mod size;
 pub use self::size::Colours as SizeColours;
@@ -26,8 +31,16 @@ mod times;
 pub use self::times::Render as TimeRender;
 // times does too
 
+#[cfg(unix)]
 mod users;
+#[cfg(unix)]
 pub use self::users::Colours as UserColours;
+#[cfg(unix)]
+pub use self::users::Render as UserRender;
 
 mod octal;
+pub use self::octal::Render as OctalPermissionsRender;
 // octal uses just one colour
+
+mod securityctx;
+pub use self::securityctx::Colours as SecurityCtxColours;
