@@ -39,7 +39,7 @@
 //! each directory)
 
 
-#[derive(PartialEq, Debug, Copy, Clone)]
+#[derive(PartialEq, Eq, Debug, Copy, Clone)]
 pub enum TreePart {
 
     /// Rightmost column, *not* the last in the directory.
@@ -253,19 +253,19 @@ mod iter_test {
     #[test]
     fn test_iteration() {
         let foos = &[ "first", "middle", "last" ];
-        let mut iter = TreeDepth::root().iterate_over(foos.into_iter());
+        let mut iter = TreeDepth::root().iterate_over(foos.iter());
 
         let next = iter.next().unwrap();
         assert_eq!(&"first", next.1);
-        assert_eq!(false, next.0.last);
+        assert!(!next.0.last);
 
         let next = iter.next().unwrap();
         assert_eq!(&"middle", next.1);
-        assert_eq!(false, next.0.last);
+        assert!(!next.0.last);
 
         let next = iter.next().unwrap();
         assert_eq!(&"last", next.1);
-        assert_eq!(true, next.0.last);
+        assert!(next.0.last);
 
         assert!(iter.next().is_none());
     }
@@ -273,7 +273,7 @@ mod iter_test {
     #[test]
     fn test_empty() {
         let nothing: &[usize] = &[];
-        let mut iter = TreeDepth::root().iterate_over(nothing.into_iter());
+        let mut iter = TreeDepth::root().iterate_over(nothing.iter());
         assert!(iter.next().is_none());
     }
 }
